Allow adding users when some have an undefined position
The position of each person in the state seems to be required, but one
of our users in state didn't have any.
The faulty user is particularly odd and seems to have been added by some
unusual means to the state. The data that existed for this user was the
following:

```
"nathan" {:away ({:from "2017-12-25", :to "2018-01-02"})}
```

While this user has a list representing the `:away` intervals, all the
others have a vector. The user also does not have any default values on
the remaining fields, and all others do.

This change makes the adding users work in the presence of that data, but we
will also remove it from the state to ensure that it doesn't break
things elsewhere.